requirements.txt
faker
pytest-playwright
allure-pytest
